Refactor: change file name upload.js to imageUpload.js
